switched to cp-sat solver for performance increase

+++ include/trace.hpp
@@ -15,7 +15,7 @@ namespace palloc {
 class Assignment {
    public:
     explicit Assignment(Coordinate dropoffCoordinate, Coordinate parkingCoordinate,
-                        uint64_t requestDuration, uint64_t routeDuration)
+                        uint32_t requestDuration, uint32_t routeDuration)
         : _dropoffCoordinate(dropoffCoordinate),
           _parkingCoordinate(parkingCoordinate),
           _requestDuration(requestDuration),
@@ -27,16 +27,16 @@ class Assignment {
     Coordinate _dropoffCoordinate;
     Coordinate _parkingCoordinate;
 
-    uint64_t _requestDuration;
-    uint64_t _routeDuration;
+    uint32_t _requestDuration;
+    uint32_t _routeDuration;
 };
 
 using Assignments = std::vector<Assignment>;
 
 class Trace {
    public:
-    explicit Trace(uint64_t timestep, uint64_t currentTimeOfDay, size_t numberOfRequests,
-                   size_t numberOfOngoingSimulations, uint64_t availableParkingSpots, double cost,
+    explicit Trace(uint32_t timestep, uint32_t currentTimeOfDay, size_t numberOfRequests,
+                   size_t numberOfOngoingSimulations, uint32_t availableParkingSpots, double cost,
                    double averageDuration, size_t droppedRequests, Assignments assignments)
         : _assignments(std::move(assignments)),
           _numberOfRequests(numberOfRequests),
@@ -52,9 +52,9 @@ class Trace {
     size_t getDroppedRequests() const noexcept;
     size_t getNumberOfRequests() const noexcept;
 
-    uint64_t getAvailableParkingSpots() const noexcept;
+    uint32_t getAvailableParkingSpots() const noexcept;
 
-    uint64_t getTimeStep() const noexcept;
+    uint32_t getTimeStep() const noexcept;
 
     double getCost() const noexcept;
     double getAverageDuration() const noexcept;
@@ -66,11 +66,11 @@ class Trace {
 
     size_t _numberOfRequests;
     size_t _numberOfOngoingSimulations;
-    size_t _availableParkingSpots;
+    uint32_t _availableParkingSpots;
     size_t _droppedRequests;
 
-    uint64_t _timestep;
-    uint64_t _currentTimeOfDay;
+    uint32_t _timestep;
+    uint32_t _currentTimeOfDay;
 
     double _cost;
     double _averageDuration;
